services/npc-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
pgvector==0.2.4
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.25.2
jsonschema==4.19.2
aiohttp==3.9.1
pyjwt==2.8.0
requests==2.31.0